requirements.txt
streamlit
pandas
numpy
geopy
scikit-learn
xgboost
matplotlib
seaborn
joblib
PyYAML
-e .